Fix for 8142485, Jake doesn't build on Windows after recent change to libjimage Reviewed-by: mchung, jlaskey

+++ src/java.base/share/classes/jdk/internal/jimage/decompressor/StringSharingDecompressor.java
@@ -224,7 +224,7 @@ public class StringSharingDecompressor implements ResourceDecompressor {
 
     @Override
     public byte[] decompress(StringsProvider reader, byte[] content,
-            int offset, int originalSize) throws Exception {
+            int offset, long originalSize) throws Exception {
         return normalize(reader, content, offset);
     }
 }
